extract wave read routes
